[EJIT] Complete the uint32_t cell-index ABI widening and harden the remaining review gaps
Address PR #143 review (ChenRuan):

1. Blocker: ejit_activate/deactivate were widened to uint32_t but the
   sibling lifecycle/cache entry points still truncated at the ABI
   boundary - ejit_invalidate(name, 256u) could invalidate instance 0
   and ejit_is_active(name, 256u) could report instance 0's state.
   Widen ejit_is_active / ejit_invalidate and the internal
   EJit::isActive / EJit::invalidateByPeriod paths to uint32_t, reject
   cellIdx >= kEJitMaxInstances before any narrowing, and make the one
   remaining narrowing (into the private uint8_t-keyed state) an
   explicit, provably-safe static_cast. DynamicCellIdxBoundaries now
   proves 256 neither queries nor invalidates instance 0.

2. EJIT_ICACHE_DIM_SIZE was consumed by LESS / math(EXPR) without the
   new numeric guard, so a non-decimal value still bypassed the uniform
   error path. It now calls ejit_require_numeric before the
   bounds/pow2 checks (negative configure test added).

3. The timing-sentinel check compared against every status value that
   exists today, so a future status assigned 0xFE would still compile
   and collide. The reservation is now structural: 0xFE is a member of
   the authoritative ejit_status_t enum itself
   (EJIT_STATUS_ICACHE_HIT_SENTINEL) - assigning it to a real status is
   a duplicate-enumerator compile error - and a static_assert locks
   kEJitIcacheHitTimingStatus to that enumerator. The per-value
   comparisons remain as an explicit second lock.

+++ llvm/include/llvm/ExecutionEngine/EJIT/EJit.h
@@ -48,7 +48,9 @@ public:
 
   bool activateAll(const std::string &periodName);
   bool deactivateAll(const std::string &periodName);
-  bool isActive(const std::string &periodName, uint8_t cellIdx) const;
+  // uint32_t like activate/deactivate: an out-of-range index is rejected
+  // (false) before any narrowing, never truncated into instance 0.
+  bool isActive(const std::string &periodName, uint32_t cellIdx) const;
 
   // Compilation
   /// Pre-computed cacheKey = funcIdx(32b) | dims(4x8b). The AOT wrapper
@@ -56,7 +58,9 @@ public:
 
   // Cache management
   void clearCache();
-  void invalidateByPeriod(const std::string &periodName, uint8_t cellIdx);
+  // uint32_t like activate/deactivate: an out-of-range index is rejected
+  // (no-op) before any narrowing, never truncated into instance 0.
+  void invalidateByPeriod(const std::string &periodName, uint32_t cellIdx);
   void invalidateAllByPeriod(const std::string &periodName);
 
   // Configuration
